Add logformats
