+++ src/Orangebeard.Client/Abstractions/Models/LogFormat.cs
@@ -1,0 +1,12 @@
+﻿namespace Orangebeard.Client.Abstractions.Models
+{
+    /// <summary>
+    /// Describes the supported log item formatting options.
+    /// </summary>
+    public enum LogFormat
+    {
+        PLAIN_TEXT,
+        HTML,
+        MARKDOWN
+    }
+}
