+++ src/Orangebeard.Client/Extensibility/TestContext.cs
@@ -1,0 +1,131 @@
+﻿using Orangebeard.Shared.Execution.Logging;
+using Orangebeard.Shared.Execution.Metadata;
+using Orangebeard.Shared.Extensibility;
+#if NET45
+using System.Runtime.Remoting;
+using System.Runtime.Remoting.Messaging;
+#endif
+using System.Threading;
+
+namespace Orangebeard.Shared.Execution
+{
+    public class TestContext : ITestContext
+    {
+        private IExtensionManager _extensionManager;
+
+        private CommandsSource _commadsSource;
+
+        public TestContext(IExtensionManager extensionManager, CommandsSource commandsSource)
+        {
+            _extensionManager = extensionManager;
+            _commadsSource = commandsSource;
+            Metadata = new TestMetadataEmitter(this, _commadsSource.TestCommandsSource as TestCommandsSource);
+        }
+
+#if !NET45
+        private AsyncLocal<ILogScope> _activeLogScope = new AsyncLocal<ILogScope>();
+
+        private AsyncLocal<ILogScope> _rootLogScope = new AsyncLocal<ILogScope>();
+
+        /// <summary>
+        /// Returns current active LogScope which provides methods for logging.
+        /// </summary>
+        public ILogScope Log
+        {
+            get
+            {
+                if (_activeLogScope.Value == null)
+                {
+                    Log = RootScope;
+                }
+
+                return _activeLogScope.Value;
+            }
+            set
+            {
+                _activeLogScope.Value = value;
+            }
+        }
+
+        private ILogScope RootScope
+        {
+            get
+            {
+                if (_rootLogScope.Value == null)
+                {
+                    //TraceLogger.Info($"New log context identified, activating {typeof(RootLogScope).Name}");
+                    RootScope = new RootLogScope(this, _extensionManager, _commadsSource);
+                }
+
+                return _rootLogScope.Value;
+            }
+            set
+            {
+                _rootLogScope.Value = value;
+            }
+        }
+
+
+#else
+        private readonly string _logicalActiveDataKey = "__AsyncLocalLogScope_Active__" + Guid.NewGuid().ToString("D");
+
+        private readonly string _logicalRootDataKey = "__AsyncLocalLogScope_Root__" + Guid.NewGuid().ToString("D");
+
+        /// <summary>
+        /// Returns current active LogScope which provides methods for logging.
+        /// </summary>
+        public ILogScope Log
+        {
+            get
+            {
+                var handle = CallContext.LogicalGetData(_logicalActiveDataKey) as ObjectHandle;
+                var activeScope = handle?.Unwrap() as ILogScope;
+
+                if (activeScope == null)
+                {
+                    activeScope = RootScope;
+                    
+                    Log = activeScope;
+                }
+
+                return activeScope;
+            }
+            set
+            {
+                CallContext.LogicalSetData(_logicalActiveDataKey, new ObjectHandle(value));
+            }
+        }
+
+        private static readonly object s_lockObj = new object();
+
+        private ILogScope RootScope
+        {
+            get
+            {
+                ILogScope rootScope;
+
+                lock (s_lockObj)
+                {
+                    var handle = CallContext.LogicalGetData(_logicalRootDataKey) as ObjectHandle;
+                    rootScope = handle?.Unwrap() as ILogScope;
+
+                    if (rootScope == null)
+                    {
+                        rootScope = new RootLogScope(this, _extensionManager, _commadsSource);
+
+                        RootScope = rootScope;
+                    }
+                }
+
+                return rootScope;
+            }
+            set
+            {
+                CallContext.LogicalSetData(_logicalRootDataKey, new ObjectHandle(value));
+            }
+        }
+#endif
+
+        public ITestMetadataEmitter Metadata { get; private set; }
+    }
+}
